mediyo_core::api

Module comments

Functions

comment_continuation

Fetch the next page of comments using the continuation token from a previous CommentsPage.

comments_page

Fetch a page of comments using a continuation token from crate::model::comments::comments_token.

reply_page

Fetch replies for a comment using the replies_continuation from a top-level Comment.